+++ source/Classes/001.ResourceBase.ps1
@@ -1,21 +1,16 @@
 <#
     .SYNOPSIS
-        A class with DSC properties that are equal for all class-based resources.
+        A class with methods that are equal for all class-based resources.
 
     .DESCRIPTION
-       A class with DSC properties that are equal for all class-based resources.
+       A class with methods that are equal for all class-based resources.
 
-    .PARAMETER DnsServer
-        The host name of the Domain Name System (DNS) server, or use 'localhost'
-        for the current node. Defaults to `'localhost'`.
+    .NOTES
+        This class should not contain any DSC properties.
 #>
 
 class ResourceBase
 {
-    [DscProperty()]
-    [System.String]
-    $DnsServer = 'localhost'
-
     # Hidden property for holding localization strings
     hidden [System.Collections.Hashtable] $localizedData = @{}