+++ CHANGELOG.md
@@ -27,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     that _Microsoft DNS Server_ is required on a node targeted by a resource,
     and that the DSC resources requires the [DnsServer](https://docs.microsoft.com/en-us/powershell/module/dnsserver)
     PowerShell module ([issue #37](https://github.com/dsccommunity/xDnsServer/issues/37)).
-  - Added the base class `ResourceBase` to hold DSC properties that should be
+  - Added the base class `ResourcePropertiesBase` to hold DSC properties that should be
     inherited for all class-based resources.
 - xDNSServerClientSubnet
   - Added integration tests.